+++ src/test/suite/extension.test.ts
@@ -15,6 +15,7 @@ suite('Extension test suite', () => {
       "myserver.com",
       "1389",
       "cn=admin,dc=example,dc=org",
+      "settings",
       "foobar",
       "dc=example,dc=org",
       "0",
@@ -38,6 +39,7 @@ suite('Extension test suite', () => {
       "${host}",
       "${port}",
       "${binddn}",
+      "${pwdmode}",
       "${bindpwd}",
       "${basedn}",
       "${sizelimit}",
@@ -55,6 +57,7 @@ suite('Extension test suite', () => {
       host: "myserver.com",
       port: "1389",
       binddn: "cn=admin,dc=example,dc=org",
+      pwdmode: "settings",
       bindpwd: "foobar",
       basedn: "dc=example,dc=org",
       sizelimit: "0",
@@ -71,6 +74,7 @@ suite('Extension test suite', () => {
     assert.strictEqual("myserver.com", connection.getHost(true));
     assert.strictEqual("1389", connection.getPort(true));
     assert.strictEqual("cn=admin,dc=example,dc=org", connection.getBindDn(true));
+    assert.strictEqual("settings", connection.getPwdMode(true));
     assert.strictEqual("foobar", connection.getBindPwd(true));
     assert.strictEqual("dc=example,dc=org", connection.getBaseDn(true));
     assert.strictEqual("0", connection.getLimit(true));
@@ -80,30 +84,30 @@ suite('Extension test suite', () => {
   });
 
   test('Test binaryGUIDToTextUUID', () => {
-    assert.strictEqual("{7f613d2a-b1ed-469b-9252-a804d4310c88}", 
+    assert.strictEqual("{7f613d2a-b1ed-469b-9252-a804d4310c88}",
       utils.binaryGUIDToTextUUID(Buffer.from("Kj1hf+2xm0aSUqgE1DEMiA==", "base64")));
 
-    assert.strictEqual("{ebaf4bc1-9068-4842-a6e5-8955bff33a6f}", 
+    assert.strictEqual("{ebaf4bc1-9068-4842-a6e5-8955bff33a6f}",
       utils.binaryGUIDToTextUUID(Buffer.from("wUuv62iQQkim5YlVv/M6bw==", "base64")));
 
-    assert.strictEqual("{32b1434c-cc5f-4ec1-8afa-e14c300a9070}", 
+    assert.strictEqual("{32b1434c-cc5f-4ec1-8afa-e14c300a9070}",
       utils.binaryGUIDToTextUUID(Buffer.from("TEOxMl/MwU6K+uFMMAqQcA==", "base64")));
 
-    assert.strictEqual("{541be56a-668c-4e78-a317-7fb201d5abc0}", 
+    assert.strictEqual("{541be56a-668c-4e78-a317-7fb201d5abc0}",
       utils.binaryGUIDToTextUUID(Buffer.from("auUbVIxmeE6jF3+yAdWrwA==", "base64")));
   });
 
   test('Test binarySIDToText', () => {
-    assert.strictEqual("S-1-5-21-4169144328-3425172002-2123581430-1103", 
+    assert.strictEqual("S-1-5-21-4169144328-3425172002-2123581430-1103",
       utils.binarySIDToText(Buffer.from("AQUAAAAAAAUVAAAACBiA+CL6J8z2R5N+TwQAAA==", "base64")));
 
-    assert.strictEqual("S-1-5-21-4169144328-3425172002-2123581430-1140", 
+    assert.strictEqual("S-1-5-21-4169144328-3425172002-2123581430-1140",
       utils.binarySIDToText(Buffer.from("AQUAAAAAAAUVAAAACBiA+CL6J8z2R5N+dAQAAA==", "base64")));
 
-    assert.strictEqual("S-1-5-21-4169144328-3425172002-2123581430-1130", 
+    assert.strictEqual("S-1-5-21-4169144328-3425172002-2123581430-1130",
       utils.binarySIDToText(Buffer.from("AQUAAAAAAAUVAAAACBiA+CL6J8z2R5N+agQAAA==", "base64")));
 
-    assert.strictEqual("S-1-5-21-3316208387-3203859757-1631524618-1117", 
+    assert.strictEqual("S-1-5-21-3316208387-3203859757-1631524618-1117",
       utils.binarySIDToText(Buffer.from("AQUAAAAAAAUVAAAAA1OpxS0F974KFz9hXQQAAA==", "base64")));
   });