+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Bind passwords can now be stored encrypted in secret storage or prompted at connection time ; this is configurable via a new field labeled "Bind Password mode" on the connection edit screen ([#64](https://github.com/fengtan/ldap-explorer/issues/64))
+- Button to reveal or hide bind passwords ([#64](https://github.com/fengtan/ldap-explorer/issues/64))
 
 ## [1.4.1] - 2025-01-13
 ### Added